fix(cli): validate-config rejects unsupported --format (M13)
`alint --format sarif validate-config` silently fell through to human
output. The earlier attempt at a fix used a blanket "reject non-default
global format" gate that fired on subcommands with their own --format
default (broke export-agents-md/suggest, reverted).

The correct fix -- the one list/facts/explain already use -- is
per-handler: validate the EFFECTIVE format (clap merges the global into
the subcommand's value, so it's caught regardless of position) against
that handler's allowed set and bail. Only validate-config lacked the
gate; it now rejects any format other than human/json (exit 2). Audited
the rest: check renders all formats; export-agents-md/suggest parse into
their own OutputFormat enum (already reject sarif); fix has a documented
human fallback. So the surface unification the prior note proposed is
unnecessary.

trycmd validate-config-format-rejected; verified both global-position and
subcommand-position invocations exit 2.

The M11 + M13 additions pushed main.rs over the 2000-line self-lint
threshold, so its test module moved to src/tests.rs (Dog2-style), keeping
the dogfood green (46/46). Post-v0.13 audit tail.
